split execution inferaces
