+++ python/ray/data/_internal/execution/interfaces/common.py
@@ -1,0 +1,2 @@
+# Node id string returned by `ray.get_runtime_context().get_node_id()`.
+NodeIdStr = str
